Joystick
Implemented joystick input forwarding and fixed minor discrepancies in the code in the joystick directory. TODO: change anything with the name joystick to gamepad.

+++ platform/joystick/JoystickAxisCode.h
@@ -2,7 +2,7 @@
 
 namespace it {
 	namespace platform {
-		enum class JoystickAxisCode : size_t {
+		enum class JoystickAxisCode : unsigned int {
 			UNKNOWN,
 			LEFT_X,
 			LEFT_Y,
